LineBoard下部に直置きだった通過駅情報をWarningInfoに移動させた
